+++ docs/about/changelog.rst
@@ -46,6 +46,9 @@ Updates
   <https://github.com/CodeGra-de/CodeGra.de/pull/1481>`__. This makes it easier
   for to start using our API. Please note that we haven't migrated all routes
   just yet, if you are missing a route please let us know!
+- The Jupyter Notebook viewer now supports more output types and colors `(#1479)
+  <https://github.com/CodeGra-de/CodeGra.de/pull/1479>`__. If a Jupyter Notebook
+  contains ANSI colored output we will display these colors in all their glory.
 
 Version Mosaic.1
 -----------------